Remove _test from test package name
The CI build is reporting 0% coverage. I suspect that this is due
to the CI build pulling down the source from the aliased import
path (hhj.me/go/spelunk) and testing that source (and not the
local source). This change removes the "_test" suffix from the
test package which puts the tests in the same package as the code
being tested. This should result in "go test" testing the local
source and reporting the correct coverage value.